# opsplatform-video-manager-backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CDNProvider:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateStreamPathRequest:
    properties:
      full_path:
        maxLength: 500
        minLength: 1
        type: string
      stream_id:
        type: integer
      table_id:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - full_path
    - stream_id
    - table_id
    type: object
...
  models.OIDCSettingsResponse:
    properties:
      client_id:
        type: string
      enabled:
        type: boolean
      frontend_success_url:
        type: string
      has_client_secret:
        type: boolean
      issuer_url:
        type: string
      redirect_url:
        type: string
      scopes:
        type: string
    type: object
  models.Stream:
    properties:
      code:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      provider:
        $ref: '#/definitions/models.CDNProvider'
      provider_id:
        description: NULL means match all providers
        type: integer
      updated_at:
        type: string
    type: object
  models.StreamPath:
    properties:
      created_at:
        type: string
      full_path:
        type: string
      id:
        type: integer
      series:
        description: derived from table_id, e.g. L001 -> "L系列"
        type: string
      stream:
        $ref: '#/definitions/models.Stream'
      stream_id:
        type: integer
      table_id:
        type: string
      updated_at:
        type: string
    type: object
  models.StreamPathImportResult:
    properties:
      created:
        type: integer
      errors:
        items:
          $ref: '#/definitions/models.StreamPathImportRowError'
        type: array
      updated:
        type: integer
    type: object
  models.StreamPathImportRowError:
    properties:
      line:
        type: integer
      message:
        type: string
      table_id:
        type: string
    type: object
...
  models.UpdateOIDCSettingsRequest:
    properties:
      clear_client_secret:
        type: boolean
      client_id:
        type: string
      client_secret:
        type: string
      enabled:
        type: boolean
      frontend_success_url:
        type: string
      issuer_url:
        type: string
      redirect_url:
        type: string
      scopes:
        type: string
    type: object
  models.UpdateStreamPathRequest:
    properties:
      full_path:
        maxLength: 500
        minLength: 1
        type: string
      stream_id:
        type: integer
      table_id:
        maxLength: 255
        minLength: 1
        type: string
    required:
    - full_path
    - stream_id
    - table_id
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Video Manager API Documentation
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Video Manager API
  version: "1.0"
paths:
...
  /api/stream-paths:
    get:
      description: Retrieve all stream paths, optionally filtered by stream ID
      parameters:
      - description: Filter by stream ID
        in: query
        name: stream_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.StreamPath'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get all stream paths
      tags:
      - stream-paths
    post:
      consumes:
      - application/json
      description: Create a new stream path with the provided information
      parameters:
      - description: Stream Path information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateStreamPathRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StreamPath'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a new stream path
      tags:
      - stream-paths
  /api/stream-paths/{id}:
    delete:
      description: Delete a stream path by its ID. This will also delete associated
        video stream endpoints.
      parameters:
      - description: Stream Path ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a stream path
      tags:
      - stream-paths
    get:
      description: Retrieve a stream path by its ID
      parameters:
      - description: Stream Path ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StreamPath'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get stream path by ID
      tags:
      - stream-paths
    put:
      consumes:
      - application/json
      description: Update an existing stream path by its ID
      parameters:
      - description: Stream Path ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated stream path information
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateStreamPathRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StreamPath'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a stream path
      tags:
      - stream-paths
  /api/stream-paths/import:
    post:
      consumes:
      - multipart/form-data
      description: 'CSV columns: 桌台号(table_id), 路径(full_path), 流区域 or stream_id. UTF-8
        with optional BOM.'
      parameters:
      - description: CSV file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.StreamPathImportResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Import stream paths from CSV
      tags:
      - stream-paths
...
  /api/system-settings/oidc:
    get:
      description: Retrieve OIDC SSO configuration (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.OIDCSettingsResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get OIDC settings
      tags:
      - system-settings
    put:
      consumes:
      - application/json
      description: Update OIDC SSO configuration (admin only)
      parameters:
      - description: OIDC settings request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateOIDCSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update OIDC settings
      tags:
      - system-settings
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"